auth-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
email-validator==2.2.0
